🚚 Rename: sidebar 폴더 정리
