+++ src/app/(private)/mypage/layout.tsx
@@ -1,4 +1,4 @@
-import MyPageSidebar from '@/features/my-page/ui/sidebar/MyPageSidebar';
+import MyPageSidebar from '@/features/my-page/common/ui/sidebar/MyPageSidebar';
 import type { WithChildren } from '@/shared/types/common';
 
 type MyPageLayoutProps = WithChildren;
